💚 modify release workflow (#112)
### TL;DR

This pull request renames and refines the release workflow specifically for the backend.

### What changed?

The release workflow has been renamed to 'backend-release.yaml'. The workflow now runs only when there are changes in the 'backend/' path. The versioning has been updated to prefix with 'backend@'. Moreover, working-directory in steps is adjusted to 'backend'.

### How to test?

Monitor GitHub Actions upon push to 'main' branch tailoring changes within the 'backend/' path to test the automated backend release process.

### Why make this change?

This change helps to differentiate the automated release process specifically for the backend, providing selective deployment and versioning scheme.

---