🔖 release version 0.10.0 (#113)
### TL;DR

Update the version of PyRB from 0.9.7 to 0.10.0

### What changed?

The version in `pyproject.toml` is updated in this pull request.

### How to test?

Verify the version number in `pyproject.toml` matches with the version specified in this PR.

### Why make this change?

The version bump is likely due to new features, improvements, or bugfixes in the application that warrant a new release.

---