🔖 release version backend@0.10.4 (#130)
### TL;DR

This Pull Request bumps up the version of `pyrb` to `0.10.4` in the backend.

### What changed?

In `backend/pyproject.toml`, the version of `pyrb` is updated from `0.10.3` to `0.10.4`.

### How to test?

After pulling the changes, check if the version string in `backend/pyproject.toml` file is `0.10.4`.

### Why make this change?

This change is necessary to release the new version `0.10.4` of the backend.

---