🔖 release version backend@0.10.6 (#134)
### TL;DR

Release new patch version 0.10.6 for the backend.

### What changed?

The version of the backend in `pyproject.toml` has been updated to `0.10.6`.

### How to test?

Checkout to the release branch and run the backend locally. Make sure all routes/services are working as expected.

### Why make this change?

This change is made to release the latest features/fixes carried out in the backend.

---