💚 fix CI workflow (#135)
### TL;DR

This PR updates the backend version and addresses some CI issues.

### What changed?

The backend version has been updated. Additionally, some issues with the CI have been fixed.

### How to test?

To test this change, run the CI pipeline and ensure that all tests pass successfully.

### Why make this change?

This change is necessary to keep the backend version up-to-date and to ensure that the CI pipeline is functioning correctly.

---