💚 use absolute paths for reusable workflow (#136)
### TL;DR

This PR fixes the CI workflow.

### What changed?

The CI workflow was updated to resolve the issues that were causing the build to fail.

### How to test?

To test this change, trigger the CI workflow and ensure that the build passes successfully.

### Why make this change?

This change was necessary to ensure that the CI workflow is functioning correctly and that the build is successful.

---